add tests for exhibition-guide redirects
